% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_empirical.R
\name{hm_distance_matrix}
\alias{hm_distance_matrix}
\title{Compute pairwise haversine distances between grid cells}
\usage{
hm_distance_matrix(coords)
}
\arguments{
\item{coords}{A data.frame with \code{lat} and \code{lon} columns in
decimal degrees. Typically \code{attr(X_land, "coords")} from
[hm_filter_land_cells()].}
}
\value{
A symmetric numeric matrix with diagonal zero.
}
\description{
Returns a symmetric \eqn{N \times N} distance matrix of great-circle
distances in kilometres between all pairs of land cells (Eq. 5 of
Clavijo Mesa et al., 2026, SF paper).
}
\examples{
\dontrun{
X_land <- hm_filter_land_cells(X)
D      <- hm_distance_matrix(attr(X_land, "coords"))
range(D)
}

}
\references{
Clavijo Mesa, M.V., Broggi, M., Di Maio, F. & Zio, E. (2026).
\emph{International Journal of Disaster Risk Reduction}, 141, 106172.
}
\seealso{
[hm_empirical_correlogram()], [hm_filter_land_cells()]
}
