% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{hm_read_netcdf}
\alias{hm_read_netcdf}
\title{Read a hazard NetCDF dataset}
\usage{
hm_read_netcdf(file, var = NULL)
}
\arguments{
\item{file}{Path to a NetCDF file.}

\item{var}{Optional. Name of the hazard variable. If `NULL` and multiple variables exist,
an error is raised (safe default).}
}
\value{
An object of class `hm_hazard`.
}
\description{
Reads a NetCDF file containing a hazard variable on a spatial grid (regular or rotated)
and returns a basic `hm_hazard` object. CRS is NOT required.
}
